[feat/#157] : PropertyCard.js 파일 분리
